job apply complete

+++ src/main/java/edu/rims/pro_connect/repository/FreelancerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import edu.rims.pro_connect.constant.UserType;
 import edu.rims.pro_connect.entity.Freelancer;
 
-public interface FreelancerRepository extends JpaRepository<Freelancer, Long> {
+public interface FreelancerRepository extends JpaRepository<Freelancer, Integer> {
 
     List<Freelancer> findByUserType(UserType freelancer);
